% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDataMatrices.R
\name{createDataMatrices}
\alias{createDataMatrices}
\title{Create lists of design matrices for each subject for each longitudinal response}
\usage{
createDataMatrices(data, formulas)
}
\description{
Create lists of design matrices for each subject for each longitudinal response
}
\keyword{internal}
