% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{difference}
\alias{difference}
\title{Take and return difference between two vectors according to some criterion.}
\usage{
difference(params.old, params.new, type)
}
\description{
Take and return difference between two vectors according to some criterion.
}
\keyword{internal}
