% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_update.R
\name{E_shape.b}
\alias{E_shape.b}
\title{Conditional expectation of Gamma log-likelihood.}
\usage{
E_shape.b(shape, X, Y, Z, tau, beta, b, w, v)
}
\description{
Conditional expectation of Gamma log-likelihood.
}
\keyword{internal}
