% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMUpdate.R
\name{EMupdate}
\alias{EMupdate}
\title{EM update}
\usage{
EMupdate(
  Omega,
  family,
  X,
  Y,
  Z,
  b,
  S,
  SS,
  Fi,
  Fu,
  l0i,
  l0u,
  Delta,
  l0,
  sv,
  w,
  v,
  n,
  m,
  hessian,
  beta.inds,
  b.inds,
  K,
  q,
  beta.quad
)
}
\description{
EM update
}
\keyword{=====================}
\keyword{Find}
\keyword{Omega,}
\keyword{Sigma}
\keyword{Unpack}
\keyword{and}
\keyword{b.hat}
\keyword{internal}
\keyword{parameter}
\keyword{the}
\keyword{vector}
