% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{extractAIC.joint}
\alias{extractAIC.joint}
\title{Extract AIC from a joint model fit.}
\usage{
\method{extractAIC}{joint}(x, scale, k = 2, conditional = FALSE, ...)
}
\description{
Extract AIC from a joint model fit.
}
