\name{GMVJM-package}
\alias{GMVJM-package}
\alias{GMVJM}
\docType{package}
\title{
\packageTitle{GMVJM}
}
\description{
\packageDescription{GMVJM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GMVJM}
\packageIndices{GMVJM}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{GMVJM}

Maintainer: \packageMaintainer{GMVJM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
1+1
}
