% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GP1_pmf_scalar}
\alias{GP1_pmf_scalar}
\title{Specifically for obtaining pmf in simulations (since requisite doesn't exist in R).}
\usage{
GP1_pmf_scalar(mu, phi, Y)
}
\description{
Specifically for obtaining pmf in simulations (since requisite doesn't exist in R).
}
\keyword{internal}
