% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Hbeta}
\alias{Hbeta}
\title{Hessian matrix on fixed effects.}
\usage{
Hbeta(beta, X, Y, Z, b, sigma, family, beta_inds, K, quad, tau, w, v)
}
\description{
Hessian matrix on fixed effects.
}
\keyword{internal}
