% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Hgammazeta}
\alias{Hgammazeta}
\title{Hessian matrix for survival parameters.}
\usage{
Hgammazeta(
  gammazeta,
  b,
  Sigma,
  S,
  SS,
  Fu,
  Fi,
  haz,
  Delta,
  w,
  v,
  b_inds,
  K,
  q,
  eps
)
}
\description{
Hessian matrix for survival parameters.
}
\keyword{internal}
