% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{joint.log.lik}
\alias{joint.log.lik}
\title{Calculate joint log-likelihood, degrees of freedom, AIC and BIC of joint model fit.}
\usage{
joint.log.lik(
  coeffs,
  dmats,
  b,
  surv,
  sv,
  l0u,
  l0i,
  gamma.rep,
  beta.inds,
  b.inds,
  K,
  q,
  family
)
}
\description{
Calculate joint log-likelihood, degrees of freedom, AIC and BIC of joint model fit.
}
\references{
Henderson R, Diggle P, Dobson A. Joint modelling of longitudinal measurements and event time data.
\emph{Biostatistics} 2000; \strong{1(4)}; 465-480.
}
\keyword{internal}
