% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint.object.R
\docType{data}
\name{joint.object}
\alias{joint.object}
\title{Fitted \code{joint} object}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
joint.object
}
\value{
A list with the following components. \describe{

 \item{TODO}{TODO}
}
}
\description{
An object returned by the \code{joint} function, inheriting
  from class \code{mjoint} and representing a fitted joint model for
  multivariate longitudinal and time-to-event data. 
  Objects of this class have methods for the generic functions \code{coef}, \code{logLik},
  \code{plot}, \code{print}, \code{ranef}, \code{fixef}, \code{summary},
  \code{AIC}, \code{getVarCov}, \code{vcov}, \code{confint}, \code{sigma},
  \code{fitted}, \code{residuals}, and \code{formula}.
}
\seealso{
\code{\link{joint}}.
}
\author{
James Murray (\email{j.murray7@ncl.ac.uk}).
}
\keyword{datasets}
