% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{joint_density}
\alias{joint_density}
\title{The joint density}
\usage{
joint_density(
  b,
  Y,
  X,
  Z,
  beta,
  D,
  sigma,
  family,
  Delta,
  S,
  Fi,
  l0i,
  SS,
  Fu,
  haz,
  gamma_rep,
  zeta,
  beta_inds,
  b_inds,
  K
)
}
\description{
The joint density
}
\keyword{internal}
