% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{joint_density_ddb}
\alias{joint_density_ddb}
\title{First derivative of the joint density with respect to b.}
\usage{
joint_density_ddb(
  b,
  Y,
  X,
  Z,
  beta,
  D,
  sigma,
  family,
  Delta,
  S,
  Fi,
  l0i,
  SS,
  Fu,
  haz,
  gamma_rep,
  zeta,
  beta_inds,
  b_inds,
  K
)
}
\description{
First derivative of the joint density with respect to b.
}
\keyword{internal}
