% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{joint_density_sdb}
\alias{joint_density_sdb}
\title{Second derivative of joint density wrt b}
\usage{
joint_density_sdb(
  b,
  Y,
  X,
  Z,
  beta,
  D,
  sigma,
  family,
  Delta,
  S,
  Fi,
  l0i,
  SS,
  Fu,
  haz,
  gamma_rep,
  zeta,
  beta_inds,
  b_inds,
  K,
  eps
)
}
\description{
Second derivative of joint density wrt b
}
\keyword{internal}
