% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lambdaUpdate}
\alias{lambdaUpdate}
\title{Update to the baseline hazard}
\usage{
lambdaUpdate(
  survtimes,
  ft,
  gamma,
  gamma_rep,
  zeta,
  S,
  Sigma,
  b,
  w,
  v,
  b_inds,
  K,
  q
)
}
\description{
Update to the baseline hazard
}
\keyword{internal}
