% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{obs.emp.I}
\alias{obs.emp.I}
\title{Calculate the observed empirical information matrix \eqn{\tilde{I}}.}
\usage{
obs.emp.I(
  Omega,
  dmats,
  surv,
  sv,
  Sigma,
  SigmaSplit,
  b,
  bsplit,
  l0u,
  w,
  v,
  n,
  family,
  K,
  q,
  beta.inds,
  b.inds
)
}
\description{
Calculate the observed empirical information matrix \eqn{\tilde{I}}.
}
\keyword{(\gamma,}
\keyword{----}
\keyword{//}
\keyword{D}
\keyword{Extract}
\keyword{Longitudinal}
\keyword{Omega}
\keyword{RE}
\keyword{Survival}
\keyword{The}
\keyword{Unpack}
\keyword{\beta}
\keyword{\sigma}
\keyword{\zeta)}
\keyword{covariance}
\keyword{data}
\keyword{dispersion}
\keyword{effects,}
\keyword{fixed}
\keyword{internal}
\keyword{matrix,}
\keyword{objects}
\keyword{parameter,}
\keyword{parameters}
