% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phi_update}
\alias{phi_update}
\title{Update for dispersion parameter (sigma for genpois family)}
\usage{
phi_update(b, X, Y, Z, beta, phi, w, v, tau)
}
\description{
Update for dispersion parameter (sigma for genpois family)
}
\keyword{internal}
