% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phi_update2}
\alias{phi_update2}
\title{Update for dispersion parameter (sigma for genpois family)}
\usage{
phi_update2(b, X, Y, Z, S, beta, phi, w, v)
}
\description{
Update for dispersion parameter (sigma for genpois family)
}
\keyword{internal}
