% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.joint.R
\name{print.summary.joint}
\alias{print.summary.joint}
\title{Print object of class summary.joint.}
\usage{
\method{print}{summary.joint}(x, digits = 3, printD = FALSE, ...)
}
\description{
Print object of class summary.joint.
}
\keyword{internal}
