% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sbeta}
\alias{Sbeta}
\title{Create vector of scores on fixed effects.}
\usage{
Sbeta(beta, X, Y, Z, b, sigma, family, beta_inds, K, quad, tau, w, v)
}
\description{
Create vector of scores on fixed effects.
}
\keyword{internal}
