% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sgammazeta}
\alias{Sgammazeta}
\title{Score vector for survival parameters.}
\usage{
Sgammazeta(
  gammazeta,
  b,
  Sigma,
  S,
  SS,
  Fu,
  Fi,
  haz,
  Delta,
  w,
  v,
  b_inds,
  K,
  q,
  eps
)
}
\description{
Score vector for survival parameters.
}
\keyword{internal}
