% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_update.R
\name{shape_update}
\alias{shape_update}
\title{Compute gradient and hessian wrt shape parameter of Gamma distn.}
\usage{
shape_update(shape, X, Y, Z, tau, beta, b, w, v)
}
\description{
Compute gradient and hessian wrt shape parameter of Gamma distn.
}
\keyword{internal}
