% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.joint.R
\name{summary.joint}
\alias{summary.joint}
\title{Summary of an \code{joint} object.}
\usage{
\method{summary}{joint}(x)
}
\arguments{
\item{x}{a joint model fit by the \code{joint} function.}
}
\value{
Object of class \code{summary.joint}.
}
\description{
Generate summary of a fitted multivariate joint model.
}
\seealso{
\code{\link{joint}} and \code{\link{joint.object}}
}
\author{
James Murray \email{j.murray7@ncl.ac.uk}
}
\keyword{internal}
