% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survmod.R
\name{surv.mod}
\alias{surv.mod}
\title{Create survival data objects based on random effects formula(e), a ph fit,
the survival data and an initial estimation of \eqn{\lambda_0}.}
\usage{
surv.mod(ph, survdata, formulas, l0.init)
}
\description{
Create survival data objects based on random effects formula(e), a ph fit,
the survival data and an initial estimation of \eqn{\lambda_0}.
}
\keyword{internal}
