% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inits.R
\name{TimeVarCox}
\alias{TimeVarCox}
\title{Obtain initial conditions for survival submodel.}
\usage{
TimeVarCox(data, b, ph, formulas, b.inds)
}
\description{
Obtain initial conditions for survival submodel.
}
\keyword{internal}
