% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vare_update}
\alias{vare_update}
\title{Update for residual variance (sigma for gaussian family)}
\usage{
vare_update(X, Y, Z, b, beta, tau, w, v)
}
\description{
Update for residual variance (sigma for gaussian family)
}
\keyword{internal}
