% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.joint}
\alias{vcov.joint}
\title{Extract the variance-covariance matrix from a \code{joint} fit.}
\usage{
\method{vcov}{joint}(x, corr = F)
}
\arguments{
\item{x}{a joint model fit by the \code{joint} function.}

\item{corr}{should the correlation matrix be returned instead of the variance-covariance?}
}
\description{
Extract the variance-covariance matrix from a \code{joint} fit.
}
\details{
Uses the observed-empirical \strong{approximation} of information matrix 
(Mclachlan & Krishnan, 2008). The estimates for the baseline hazard are not estimated.
}
\references{
McLachlan GJ, Krishnan T. \emph{The EM Algorithm and Extensions.} Second Edition. 
Wiley-Interscience; 2008.
}
\seealso{
\code{\link{obs.emp.I}}
}
\author{
James Murray \email{j.murray7@ncl.ac.uk}
}
\keyword{internal}
