% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{bind.bs}
\alias{bind.bs}
\title{create appropriately-dimensioned matrix of random effects.}
\usage{
bind.bs(bsplit)
}
\description{
create appropriately-dimensioned matrix of random effects.
}
\keyword{internal}
