% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMUpdate.R
\name{EMupdate}
\alias{EMupdate}
\title{EM Update}
\usage{
EMupdate(
  Omega,
  family,
  X,
  Y,
  Z,
  b,
  S,
  SS,
  Fi,
  Fu,
  l0i,
  l0u,
  Delta,
  l0,
  sv,
  w,
  v,
  n,
  m,
  hessian,
  beta.inds,
  b.inds,
  K,
  q,
  beta.quad
)
}
\description{
EM Update
}
\keyword{internal}
