% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{extractAIC.joint}
\alias{extractAIC.joint}
\title{Extract AIC from a joint model fit.}
\usage{
\method{extractAIC}{joint}(fit, scale, k = 2, conditional = FALSE, ...)
}
\arguments{
\item{fit}{A fitted \code{joint} object,}

\item{scale}{See \code{\link[stats]{extractAIC}}; not used.}

\item{k}{Numeric specifying the "weight" of degrees of freedom (default \code{k=2}).}

\item{conditional}{Should AIC of conditional or observed log-likelihood be used? Defaults
to \code{conditional = FALSE}.}

\item{...}{additional arguments (none used).}
}
\description{
Extract AIC from a joint model fit.
}
