% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{joint.log.lik}
\alias{joint.log.lik}
\title{Calculate log-likelihood and degrees of freedom from a multivariate joint model}
\usage{
joint.log.lik(
  coeffs,
  dmats,
  b,
  surv,
  sv,
  l0u,
  l0i,
  gamma.rep,
  beta.inds,
  b.inds,
  K,
  q,
  family,
  Sigma
)
}
\description{
Calculate log-likelihood and degrees of freedom from a multivariate joint model
}
\seealso{
\code{\link{logLik.joint}}
}
\author{
James Murray (\email{j.murray7@ncl.ac.uk})
}
\keyword{internal}
