% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{obs.emp.I}
\alias{obs.emp.I}
\title{Calculate the observed empirical information matrix \eqn{\tilde{I}}.}
\usage{
obs.emp.I(
  Omega,
  dmats,
  surv,
  sv,
  Sigma,
  SigmaSplit,
  b,
  bsplit,
  l0u,
  w,
  v,
  n,
  family,
  K,
  q,
  beta.inds,
  b.inds,
  beta.quad
)
}
\description{
Calculate the observed empirical information matrix \eqn{\tilde{I}}.
}
\keyword{internal}
