% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynPred.R
\name{plot.dynPred}
\alias{plot.dynPred}
\title{Plot conditional survival probabilities.}
\usage{
\method{plot}{dynPred}(x, what = c("median", "mean"), ...)
}
\arguments{
\item{x}{an object with class \code{dynPred}.}

\item{what}{should the \code{"median"} (the default) or \code{"mean"} probability be displayed?}

\item{...}{additional arguments (none used).}
}
\description{
Produces a simple plot of the probability the subject survives given failure
times along with the 95\% confidence interval.
}
\seealso{
\code{\link{dynPred}}
}
\author{
James Murray (\email{j.murray7@ncl.ac.uk}).
}
\keyword{internal}
