% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{plot.residuals.joint}
\alias{plot.residuals.joint}
\title{Plot joint model residuals}
\usage{
\method{plot}{residuals.joint}(x, ...)
}
\arguments{
\item{x}{an object with class \code{residuals.joint}.}

\item{...}{additional arguments (none used).}
}
\description{
Produces a simple (panelled) plot of the residuals obtained by a
joint model (obtained by \code{\link{joint}}) of the same type as the original response.
}
\seealso{
\code{\link{residuals.joint}}
}
\author{
James Murray (\email{j.murray7@ncl.ac.uk}).
}
