% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.joint.R
\name{summary.joint}
\alias{summary.joint}
\title{Summary of an \code{joint} object.}
\usage{
\method{summary}{joint}(object, ...)
}
\arguments{
\item{object}{a joint model fit by the \code{joint} function.}

\item{...}{additional arguments (none used).}
}
\value{
Object of class \code{summary.joint}.
}
\description{
Generate summary of a fitted multivariate joint model.
}
\examples{
# Simple univariate on log(serum bilirubin) ----------------------------
data(PBC)
long.formulas <-  list(
  log(serBilir) ~ drug * (time + I(time^2)) + (1 + time + I(time^2)|id)
)
surv.formula <- Surv(survtime, status) ~ sex + drug
fit <- joint(long.formulas = long.formulas,
             surv.formula = surv.formula,
             data = PBC, family = list("gaussian"))
summary(fit)              

# Bivariate with a dispersion model ------------------------------------
\donttest{
PBC <- na.omit(PBC[,c('id', 'survtime', 'status', 'sex', 
                      'drug', 'platelets', 'albumin', 'time')])
long.formula <- list(
  platelets ~ time * drug + (1 + time|id),
  albumin ~ time * drug + (1 + time|id)
)
surv.formula <- Surv(survtime, status) ~ sex + drug
fit <- joint(long.formula, surv.formula, PBC, 
             family = list("negbin", "gaussian"),
             disp.formula = list(~time, ~1))
summary(fit)
}
}
\seealso{
\code{\link{joint}} and \code{\link{joint.object}}
}
\author{
James Murray \email{j.murray7@ncl.ac.uk}
}
