% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2xtab.R
\name{xtable.joint}
\alias{xtable.joint}
\title{Print an LaTeX-ready \code{xtable} for a \code{joint} object.}
\usage{
\method{xtable}{joint}(
  x,
  p.val = FALSE,
  caption = TRUE,
  max.row = NULL,
  dp = 3,
  capture = FALSE,
  capture.location = "",
  hlines = "middle-bottom",
  booktabs = TRUE,
  size = "footnotesize",
  ...
)
}
\arguments{
\item{x}{a joint model fit by the \code{joint} function.}

\item{p.val}{logical, should p-values be returned? Defaults to \code{p.val = FALSE}.}

\item{caption}{logical, should a generic caption be attached to the printed table?
Defaults to \code{caption = TRUE}, which prints out some computation details.}

\item{max.row}{integer, the number of rows after which the table is `broken' vertically
and merged horizontally; useful for long tables. Defaults to \code{max.row = NULL} which
results in one long table. Note that this can be quite finicky, so trial and error may be 
required.}

\item{dp}{integer, the number of decimal places to round the estimate, standard error and
confidence intervals to; defaults to \code{dp = 3}.}

\item{capture}{logical, should the printed \code{xtable} output be saved anywhere instead
of just printed to the console? Defaults to \code{capture = FALSE}.}

\item{capture.location}{character, if \code{capture = TRUE}, this should specify what
file it should be saved in. Defaults to \code{capture.location = ""}.}

\item{hlines}{character, specifies which horizontal lines are used in the outputted
LaTeX table. Supply a character string which contains \code{"top"}, \code{"middle"} and/or
\code{"bottom"} (in any order) to specify a \code{toprule}; \code{midrule} and 
\code{bottomrule} in the table. If \code{booktabs = FALSE}, then these will simply 
be \code{hline}s. For instance \code{hlines = "top-middle-bottom"} prints all three;
whilst \code{hlines = "middle-bottom"} skips the \code{toprule}.}

\item{booktabs}{logical, if \code{booktabs = TRUE} (the default) then\code{toprule};
\code{midrule} and \code{bottomrule} replace the usual \code{hline}s.}

\item{size}{character, LaTeX size to be placed before the tabular environment, defaults
to \code{size = "footnotesize"}; replace with \code{"normalsize"} if wanted.}

\item{...}{additional arguments, none used.}
}
\value{
A LaTeX-ready \code{xtable} print-out of the joint model.
}
\description{
Prints an \code{xtable} output for a fitted \code{joint} object to the console,
or to a specified save location
}
\examples{
\donttest{
# Bivariate joint model ------------------------------------------------
require(xtable)
data <- simData(n = 100)$data
long.formula <- list(
  Y.1 ~ time + cont + bin + (1 + time|id),
  Y.2 ~ time + cont + bin + (1 + time|id)
)
surv.formula <- Surv(survtime, status) ~ cont + bin
family <- list("gaussian", "gaussian")
fit <- joint(long.formula, surv.formula, data, family)
xtable(fit)
# Example of arguments: remove caption, add p-values.
xtable(fit, p.val = TRUE, dp = 4, caption = FALSE)
# Change size, place horizontal lines everywhere
xtable(fit, size = "normalsize", hlines = c("top-middle-bottom"))
# Make a wider table without booktabs 
xtable(fit, booktabs = FALSE, max.row = 6)
}
}
\seealso{
\code{\link{joint}}
}
\author{
James Murray (\email{j.murray7@ncl.ac.uk}).
}
